requirements.txt
langchain-text-splitters
langchain-google-vertexai
langchain-openai